requirements.txt
Flask==2.3.2
Werkzeug==2.3.2
Jinja2==3.1.2
MarkupSafe==2.1.1
click==8.1.3
itsdangerous==2.1.2
torch==1.12.0
torchvision==0.13.0
numpy==1.22.0
Pillow==9.0.1
opencv-python==4.6.0.66
olefile==0.46
scipy==1.8.0
typing-extensions==4.1.1
